application-generator/src/content/python/modules/home/requirements.txt
alembic==1.8.0
click==8.1.3
colorama==0.4.5
distlib==0.3.4
dominate==2.6.0
filelock==3.6.0
Flask==2.1.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.10
Flask-DotEnv==0.1.2
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
platformdirs==2.5.2
python-dotenv==0.20.0
six==1.16.0
SQLAlchemy==1.4.40
virtualenv==20.14.1
visitor==0.1.3
Werkzeug==2.1.2
